Internal: rename `flatten` to `flat_graph`
(preparation for EP-3698 improved UDP support and docs)

`flatten` as active verb sound too much like an in-place
operation on the object and does not correspond with the
"exporting" nature of the method.
`flat_graph` is also bit more explicit that it is about
flat dict/map representation of a process graph
and less likely to collide with other kinds of flattening.